+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Logger for GHEtool (issue #96).
 - Examples are now also in RTD.
 - Reynolds number is shown on the result page (issue #112).
-- It is now possible to use building loads instead of ground loads (issue #115).
+- It is now possible to use building loads (with a SCOP/SEER) instead of ground loads(issue #115).
 
 ### Changed
 - In figure plotting, the interval[x[i], x[i+1]) now has the value y[i] (instead of y[i-1]).